Big one: I wish this were dogfood
Yes, thats a big one and I am not proud of it. Before I explain a bit to
observant reader what I did here, I want to excuse myself for not taking
the time to cleanup the commit history.

Lets start with the ultimate goal of these changes: enable the use of
multiple chips on the same machine with this tool. This required
refactoring of certain parts in the code (we might have killed some bugs
with that as well) and more sophisticated usage of the bindings. Devices
can now be queried and selected specifically, however if only one
device is connected, this is not necessary.

Second error handling is improved. More wrappers are implemented in
rust_mcd::raw; furthermore these wrappers do not return anyhow::Result
anymore but the exact code returned from the API call so that calling
functions can actually check the return type and maybe call the function
again if desired. rust_mcd::raw should remain as simple as possible.

The linux setup for the docker probably compiles, but is completely
untested.